STAT: Added tolerance for redundant arguments
Till now the tool was failing if some of the redundant
arguments were specified in the command line. For instance,
if one uses '-a' option when only one product configuration
is present.
Now in case of a redundant argument STAT simply ignores it
and only prints at the end a warning note with the details.

In addition, enhanced naming for project files generated
for STAT IDE.

Signed-off-by: Arseniy Aharonov <Arseniy.Aharonov@wdc.com>