STAT: Fixed the link issue of an error-message within debugger of MSVS IDE
The MSVS IDE always assumes that $(ProjectDir) and CWD from which
NMAKE was called are the same directory. However, keeping project files
only within the root directory of the framework is a very harsh limitation
that produces many temporary files within framework's directory, thus
obscuring the image of the contents within it.

Signed-off-by: Arseniy Aharonov <arseniy@aharonov.icu>