STAT: Refactor build-tools structure to support multiple build tools
As part of changing STAT to support build-tools, it undergoes
refactoring that will ensure easier adding of new build-tools
and enable selection of correct build-tools based on different factors
(e.g. OS).
All unit-tests are passing now on Linux, however without the very support for
Linux as an environment to run the STAT framework.

Signed-off-by: Arseniy Aharonov <arseniy@aharonov.icu>