MAKE: Added a GNU Make-tool executable to further replace NMAKE usage
GNU Make-tool has a more robust set of features and a much flexible
and convenient syntax, not to mention is more stable.
GNU Make-tool also has cross-platform support, which allows using
it on more systems running different OS-s, and keep the same
format of makefiles.
Linux has this tool quite accessible along with GCC, whereas in
Windows it's that straightforward. Therefore, GNU Make-tool was
compiled for Windows and added to this repo, to ensure its
accessibility on Windows-based platform.

Signed-off-by: Arseniy Aharonv <arseniy@aharonov.icu>