STAT: Changed ability to run an ignored makefile if specified explicitly
If a makefile was added to `.statignore` file there was no way to run it.
Now if such file is explicitly specified in the command-line, it will
bypass the stat-ignore filtering.
This behavior allows the developers to directly run broken test-packages,
if the latter are registered in the stat-ignore list.